% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcelStats_cat_df.R
\name{ExcelStats_cat_df}
\alias{ExcelStats_cat_df}
\title{Calculate Descriptive Statistics for a Categorical Variable as a Data Frame}
\usage{
ExcelStats_cat_df(df, categorical_var, value_var)
}
\arguments{
\item{df}{A data frame containing the variables.}

\item{categorical_var}{The name of the categorical variable.}

\item{value_var}{The name of the value variable.}
}
\value{
A data frame with descriptive statistics for each category.
}
\description{
This function calculates descriptive statistics for a categorical variable in a data frame and returns them as a data frame.
}
\examples{
set.seed(666)
library(tibble)
df <- tibble(Kjønn = sample(c("Kvinne", "Mann"), 150, replace = TRUE), sofrwrk = rnorm(150))
ExcelStats_cat_df(df, "Kjønn", "sofrwrk")
}
