% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcelStats_df.R
\name{ExcelStats_df}
\alias{ExcelStats_df}
\title{Calculate Descriptive Statistics for Multiple Variables as a Data Frame}
\usage{
ExcelStats_df(df, ...)
}
\arguments{
\item{df}{A data frame containing the variables.}

\item{...}{Unquoted variable names to calculate statistics for.}
}
\value{
A data frame with descriptive statistics for each variable.
}
\description{
This function calculates descriptive statistics for given variables in a data frame and returns them as a data frame.
}
\examples{
set.seed(666)
df <- tibble(x = sample(1:10, 150, replace = TRUE), y = sample(18:67, 150, replace = TRUE))
df \%>\% ExcelStats_df(x, y)
}
