% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcelStats_kbl.R
\name{ExcelStats_kbl}
\alias{ExcelStats_kbl}
\title{Calculate Descriptive Statistics for Multiple Variables as a Kable}
\usage{
ExcelStats_kbl(df, ..., digits = 2)
}
\arguments{
\item{df}{A data frame containing the variables.}

\item{...}{Unquoted variable names to calculate statistics for.}
}
\value{
A kable object with descriptive statistics for each variable.
}
\description{
This function calculates descriptive statistics for given variables in a data frame and returns them as a kable object for easy printing.
}
\examples{
library(tibble)
library(ExcelStats)
set.seed(666)
df <- tibble(x = sample(1:10, 150, replace = TRUE), y = sample(18:67, 150, replace = TRUE))
df \%>\% ExcelStats_kbl(x, y, digits = 2)
}
