% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcelStats_df.R
\name{ExcelStats_df}
\alias{ExcelStats_df}
\title{Calculate Descriptive Statistics as a Data Frame}
\usage{
ExcelStats_df(df, variable)
}
\arguments{
\item{df}{A data frame containing the variable.}

\item{variable}{The name of the variable to calculate statistics for.}
}
\value{
A data frame with descriptive statistics.
}
\description{
This function calculates descriptive statistics for a given variable in a data frame and returns them as a data frame.
}
\examples{
set.seed(666)
library(tibble)
df <- tibble(x = sample(1:10, 150, replace = TRUE))
ExcelStats_df(df, "x")
}
