% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcelStats_kbl.R
\name{ExcelStats_kbl}
\alias{ExcelStats_kbl}
\title{Calculate Descriptive Statistics as a Kable}
\usage{
ExcelStats_kbl(df, variable)
}
\arguments{
\item{df}{A data frame containing the variable.}

\item{variable}{The name of the variable to calculate statistics for.}
}
\value{
A kable object with descriptive statistics.
}
\description{
This function calculates descriptive statistics for a given variable in a data frame and returns them as a kable object for easy printing.
}
\examples{
set.seed(666)
df <- tibble(x = sample(1:10, 150, replace = TRUE))
ExcelStats_kbl(df, "x")
}
